move blitz babel config into nextjs fork
