+++ nextjs/packages/next/package.json
@@ -84,6 +84,7 @@
     "ast-types": "0.13.2",
     "b64-lite": "^1.4.0",
     "bad-behavior": "^1.0.1",
+    "babel-plugin-superjson-next": "0.4.2",
     "browserify-zlib": "0.2.0",
     "browserslist": "4.16.6",
     "buffer": "5.6.0",
