+++ nextjs/test/unit/fix-node-file-trace/fix-node-file-trace.unit.test.ts
@@ -1,0 +1,9 @@
+import pluginTester from 'babel-plugin-tester'
+import path from 'path'
+import FixNodeFileTrace from 'next/dist/build/babel/plugins/fix-node-file-trace'
+
+pluginTester({
+  pluginName: FixNodeFileTrace.name,
+  plugin: FixNodeFileTrace,
+  fixtures: path.join(__dirname, 'tests'),
+})
