+++ nextjs/test/unit/fix-node-file-trace/tests/pages/api/api-route/code.js
@@ -1,0 +1,3 @@
+export default function HealthCheck(_req, res) {
+  res.status(200).send('ok')
+}
