+++ nextjs/test/unit/fix-node-file-trace/tests/pages/gSP arrow function with implicit return/code.js
@@ -1,0 +1,7 @@
+export const getStaticProps = () => ({
+  props: { today: new Date() },
+})
+
+export default function IndexPage({ today }) {
+  return JSON.stringify({ today })
+}
