+++ nextjs/test/unit/fix-node-file-trace/tests/pages/gSSP function declaration/code.js
@@ -1,0 +1,17 @@
+export async function getServerSideProps() {
+  const products = [
+    {
+      name: 'Hat',
+      publishedAt: new Date(0),
+    },
+  ]
+  return {
+    props: {
+      products,
+    },
+  }
+}
+
+export default function Page({ products }) {
+  return JSON.stringify(products)
+}
