+++ nextjs/test/unit/fix-node-file-trace/tests/pages/separate export declaration/code.js
@@ -1,0 +1,20 @@
+export const getServerSideProps = async () => {
+  const products = [
+    {
+      name: 'Hat',
+      publishedAt: new Date(0),
+    },
+  ]
+
+  return {
+    props: {
+      products,
+    },
+  }
+}
+
+function Page({ products }) {
+  return JSON.stringify(products)
+}
+
+export default Page
