--- packages/babel-preset/README.md
@@ -1,19 +1,0 @@
-[![npm version](https://badge.fury.io/js/@blitzjs/babel-preset.svg)](https://badge.fury.io/js/@blitzjs/babel-preset)
-
-# @blitzjs/babel-preset
-
-## Setup
-
-1. Install it using npm: `npm install -D @blitzjs/babel-preset`.
-2. Add `blitz` preset in your `babel.config.js`.
-
-Example `babel.config.js`:
-
-```js
-module.exports = {
-  presets: ['@blitzjs'],
-  presets: [],
-};
-```
-
-3. You're done!
