--- packages/babel-preset/jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  preset: '../../jest-unit.config.js',
-  testEnvironment: 'jest-environment-jsdom',
-};
