--- packages/babel-preset/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@blitzjs/babel-preset",
-  "version": "0.41.2-canary.4",
-  "license": "MIT",
-  "scripts": {
-    "test": "jest",
-    "test:watch": "jest --watch"
-  },
-  "module": "dist/blitzjs-babel-preset.esm.js",
-  "main": "dist/blitzjs-babel-preset.cjs.js",
-  "typings": "dist/blitzjs-babel-preset.cjs.d.ts",
-  "author": "Brandon Bayer",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "prettier": {
-    "printWidth": 80,
-    "semi": true,
-    "singleQuote": true,
-    "trailingComma": "es5"
-  },
-  "size-limit": [
-    {
-      "path": "dist/babel-preset.cjs.production.min.js",
-      "limit": "10 KB"
-    },
-    {
-      "path": "dist/babel-preset.esm.js",
-      "limit": "10 KB"
-    }
-  ],
-  "dependencies": {
-    "@babel/helper-module-imports": "^7.0.0",
-    "babel-plugin-superjson-next": "0.3.0"
-  },
-  "devDependencies": {
-    "@babel/core": "7.12.10"
-  }
-}
