--- packages/babel-preset/src/fix-node-file-trace/fix-node-file-trace.test.ts
@@ -1,9 +1,0 @@
-import pluginTester from 'babel-plugin-tester';
-import path from 'path';
-import FixNodeFileTrace from './index';
-
-pluginTester({
-  pluginName: FixNodeFileTrace.name,
-  plugin: FixNodeFileTrace,
-  fixtures: path.join(__dirname, 'tests'),
-});
