--- packages/babel-preset/src/fix-node-file-trace/tests/pages/api/api-route/code.js
@@ -1,3 +1,0 @@
-export default function HealthCheck(_req, res) {
-  res.status(200).send('ok');
-}
