--- packages/babel-preset/src/fix-node-file-trace/tests/pages/api/api-route/output.js
@@ -1,7 +1,0 @@
-import { withFixNodeFileTrace as _withFixNodeFileTrace } from 'next/dist/server/utils';
-
-function HealthCheck(_req, res) {
-  res.status(200).send('ok');
-}
-
-export default _withFixNodeFileTrace(HealthCheck);
