--- packages/babel-preset/src/fix-node-file-trace/tests/pages/gSP arrow function with implicit return/output.js
@@ -1,11 +1,0 @@
-import { withFixNodeFileTrace as _withFixNodeFileTrace } from 'next/dist/server/utils';
-export const getStaticProps = _withFixNodeFileTrace(() => ({
-  props: {
-    today: new Date(),
-  },
-}));
-export default function IndexPage({ today }) {
-  return JSON.stringify({
-    today,
-  });
-}
