--- packages/babel-preset/src/fix-node-file-trace/tests/pages/gSP arrow function/code.js
@@ -1,17 +1,0 @@
-export const getStaticProps = async () => {
-  const products = [
-    {
-      name: 'Hat',
-      publishedAt: new Date(0),
-    },
-  ];
-  return {
-    props: {
-      products,
-    },
-  };
-};
-
-export default function Page({ products }) {
-  return JSON.stringify(products);
-}
