--- packages/babel-preset/src/fix-node-file-trace/tests/pages/transforms a valid example/code.js
@@ -1,18 +1,0 @@
-export const getServerSideProps = async () => {
-  const products = [
-    {
-      name: 'Hat',
-      publishedAt: new Date(0),
-    },
-  ];
-
-  return {
-    props: {
-      products,
-    },
-  };
-};
-
-export default function Page({ products }) {
-  return JSON.stringify(products);
-}
