--- packages/babel-preset/src/types.d.ts
@@ -1,9 +1,0 @@
-declare module '@babel/helper-module-imports' {
-  import { NodePath, types } from '@babel/core';
-
-  function addNamed(
-    path: NodePath,
-    named: string,
-    source: string
-  ): types.Identifier;
-}
